Increase Test Coverage
